scraper/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
playwright==1.49.1
supabase==2.11.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
httpx==0.28.1
pydantic==2.10.4
